+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features/Changes
 
+- [#1653](https://github.com/lapce/lapce/pull/1653): Paint plugin icons
 - [#1644](https://github.com/lapce/lapce/pull/1644): Added "Reveal in File Tree" action to the editor tabs context menu
 - [#1645](https://github.com/lapce/lapce/pull/1645): Add plugin search